OSCAR-869: Units/Time field on billing can now accept numbers with decimals

+++ src/main/java/oscar/oscarBilling/ca/on/data/BillingClaimHeader1Data.java
@@ -68,7 +68,7 @@ public class BillingClaimHeader1Data {
 	double cashTotal;
 	double debitTotal;
 	String providerName;
-	int numItems;
+	float numItems;
 	
 	public String getSer_num() {
 		return ser_num;
@@ -496,11 +496,11 @@ public class BillingClaimHeader1Data {
 		this.provider_no = provider_no;
 	}
 
-	public int getNumItems() {
+	public float getNumItems() {
 		return numItems;
 	}
 
-	public void setNumItems(int numItems) {
+	public void setNumItems(float numItems) {
 		this.numItems = numItems;
 	}
